phpunit bootstrap: Allow overriding $hasIntegrationTests
Some of Wikibase’s self-proclaimed unit tests don’t actually work when
run as pure unit tests, and as long as the unit-ness isn’t enforced in
CI, I doubt that’s going to change – but it makes running the tests on
their own annoying (I noticed this again in T362505). Let’s add a way to
override the detection via an environment variable.

Change-Id: I9aae6b2446d0471d44474654b57a2d48849a4d88