jquery.makeCollapsible: Make toggle buttons work inside cloned content
Collapsible elements may be cloned, e.g. for use in reference popups (see T362561). In this case they should keep their functionality.

This patch adds an instruction that makes sure the button element is inside the collapsible element (which it should be when no custom togglers are used, if my understanding is correct). If not, it looks for the closest `.mw-collapsible` element up in the DOM tree.

If I understand correctly, the first `actionHandler` declaration relates to the case when no custom togglers is used, so changing it won't affect the behavior with custom togglers which would require a different approach.

Bug: T362561
Change-Id: I1f8f1108a17246f16f89599025cb128ede307161