Don't use content language in Message::getTitle()
Change the Message::getTitle() function to not use the
wiki's content language when generating links to pages in the
MediaWiki namespace.

Message::getTitle() returns the uppercased version of the message
name as a title object. For some languages, such as Azeri and Turkish,
this can lead to some transformations we don't want. For instance,
when you uppercase "i" in Azerbaijani, it becomes "İ" (dotted I),
leading to links to "MediaWiki:İpbreason-dropdown" on Special:Block.

Bug: T362654
Change-Id: Id37c226270c53352760790ddd461ca98920b5910