[Parser] Add deprecated Parser:$proofreadRenderingPages property
We are deprecating dynamic properties on the Parser object, but as a first
step we're making them *actual* properties, so that we can draw a line in
the sand against adding new dynamic properties.  In a follow up we'll
actually remove this property from the ProofreadPage extension.

Also add an entry to RELEASE-NOTES-1.42, which
Ifcb9492cc5c814d702c6e61e8231abfd8ea0647c forgot to include.

Bug: T362664
Bug: T343227
Followup-To: Ifcb9492cc5c814d702c6e61e8231abfd8ea0647c
Change-Id: Id3dbc99d92f2769d4e5c95293cf1a277b98ddd3b