Reduce FOUC/reflows when using .mw-collapsible-toggle-placeholder
Improve the CSS that temporarily shows non-clickable toggles for
collapsible elements (until the JS runs and adds the real toggles)
to support toggle placeholders (.mw-collapsible-toggle-placeholder)
and premade toggles (.mw-collapsible-toggle).

We can do this decently these days using the `:has()` selector,
available in the most recent versions of all browsers.

Note that this is not intended to handle every case, especially
with nested collapsibles, as that's still impossible in just CSS.

This can be tested using a copy of the advanced demo page:
https://mediawiki.org/wiki/Manual:Collapsible_elements/Demo/Advanced
by looking at the "Pre-made toggle" and "Custom toggle link placement"
examples.

Change-Id: I91fd19bb71f7104dbe856c0a1a43b70c6e6c44fd