[Parser] Add deprecated Parser::$mTemplatePath property
We are deprecating dynamic properties on the Parser object, but as a
first step we're making them *actual* properties, so that we can draw
a line in the sand against adding new dynamic properties.  In a follow
up we'll actually remove this property from the
LabeledSectionTransclusion extension.

An entry in RELEASE-NOTES-1.42 was already added in
Id3dbc99d92f2769d4e5c95293cf1a277b98ddd3b.

Bug: T362693
Bug: T343227
Bug: T362692
Followup-To: I701a27459704839f9f996a171b6982ed17cdd00b
Change-Id: I3a992f4d601135b6327d941a1852e861d1e82d78