api: Pass wikiId to UserIdentity instance for action=userrights
Action API action=userrights is using the special page to lookup the
user and since c87c8f64 this returns a UserIdentityValue instead of
UserRightsProxy for cross-wiki changes
But the getId() call was not updated to use the same wikiid to get the
id. This results in an assertion error. Before that change api was
returning the foreign user id (via UserRightsProxy), so it should remain
to do that.

Bug: T362740
Change-Id: If0cd239e8603d5b577184f0122cb74a28dafdab8
(cherry picked from commit 190012b69976b3f3b57342a65ce8509f77dd46ff)