rdbms: Protect against stale cache in LB::getMaxLag()
In get lags, you get the array of server index to lag value and servers
can be depooled (which means server ids will be moved around), that
means sometimes we end up with server index higher than max in the
config leading to "PHP Notice: Undefined offset" errors.

This is not optimal, this patch basically means mediawiki will be lying
about the server that has the highest lag for ten seconds after each
depool but name of that server is not used in any meaningful way so meh.

We can't really change the output format of LB::getLagTimes() as it's
used like this and many consumer outside are relying on it and I can't
fix this on its own.

Bug: T361824
Change-Id: I0a512138c79a9ce91620cf10f495255eb8512000