Replace a few open/closeElement with Html::rawElement
The two separate methods can be problematic as there is no enforced
relationship between the two. I believe the few suggestions in this
patch are easy wins that help making the code more robust.

Intentionally not touching anything else to make the rebase of
possible conflicts as easy as possible. I know there is a lot more
that could be replaced. Let's do this in other patches.

Bug: T356544
Change-Id: I47c05314401f9b8ea236ebb006dfe9ade7600abe