ParserOutput: Temporarily write (unused) TOCHTML to ParserCache
Even though this JSON property is unused on master, the previous
train release read it from the JSON (and threw the value away).
In order to provide error-free roll-forward and roll-back of the
train, temporarily write an empty string as the value of TOCHTML
so that the read from `$jsonData['TOCHTML']` won't cause a PHP
notice in the logs if we roll back.

This patch is only needed for one train release, and can then
be removed.

Bug: T363107
Change-Id: I77add3bd7f00941cb81481f738bc59d6008c2406