block: Fix overly broad block selection in getRangeCond()
In the new schema, getRangeCond() with a specified range of say
1.2.0.0/16 would match blocks of narrower range with the same start,
e.g. 1.2.0.0/24, because a match of bt_ip_hex was sufficient, but that
is the same as the start of the range.

So, when searching for single IP blocks, add a condition requiring
bt_range_start IS NULL.

This affects the bkip parameter to ApiQueryBlocks and searches from
Special:BlockList.

Add regression test.

Change-Id: Ie3943befb8a02cc4939624f7bad48ae50056cd7d