CodexModule: Stop providing "legacy" builds of Codex styles
Previously we provided an alternate build of Codex styles under
the name "theme-wikimedia-ui-legacy" with some different values
for size/font-size. We have removed all distinctions between
legacy and non-legacy builds of Codex upstream, so this change
updates CodexModule.php on MediaWiki to stop serving the
"legacy"-specific Codex files when they are requested. Instead,
the default Codex files will be provided instead.

In the future these code paths can be removed entirely; for now
this can stay for backwards compatibility.

Bug: T360806
Change-Id: Ic114e284d34cc9ea29701085ed766484cbeea3e8