docs: document $.fn.injectSpinner() with a typedef
Previously, the documentation referenced the `@ignore`d function
createSpinner() for the type of its parameter. Move the object
description to a @typedef and copy the string description so that it
makes sense without the documentation of the `@ignore`d function.
Also make the documentation stricter by using string literals instead of
just `string`, and add the type to the defaults variable as well (this
only helps IDEs, it isn’t exposed).

Bug: T360241
Depends-On: I8afac932db9feac07c77e336eaffcc2cb34b7f4a
Change-Id: Ib3da99597004291f5c89734ec4e15e03fc0742d7