Fix tests/phpunit/bootstrap.php on Windows
If you want to run PHP code on Windows, you have to specify the
php.exe executable, not just the file containing the code.
Windows doesn't read #! lines and will run the file using whatever
executable is associated with the file extension, which might be
nothing or a text editor.

This also helps on other systems to ensure that the new process uses
the same PHP executable as the current one.

Change-Id: I666825dea7947954b9c59d934c7232a5eac208cd