Fix MaintenanceRunner test failures on Windows
In the following test cases in MaintenanceRunnerTest.php:

  yield 'extension script path, using prefix'
    => [ 'FakeExtension:fakeScript.php', FakeScript::class ];

  // NOTE: assumes the class has been loaded by the previous test case!
  yield 'extension class, using prefix'
    => [ 'FakeExtension:FakeScript', FakeScript::class ];

…the noted assumption did not quite hold on Windows.

After including …/fakeScript.php, the class was indeed loaded. The
next case would use the loaded class, but before it did that, it
first tries to include …/FakeScript.php - which doesn't exist, but
it's the same as …/fakeScript.php on case-insensitive filesystems,
so the file was loaded twice, causing "Fatal error: Cannot declare
class MediaWiki\Extension\FakeExtension\Maintenance\FakeScript,
because the name is already in use in …\fakeScript.php on line 0".

Re-order some code in MaintenanceRunner, so that it tries to use the
existing class (or autoload it) before it tries to include the file.

Change-Id: Ic7ed7139bbede48097df59338c82688081688c3b