Add months as duration to Language:formatDuration
I think there is further room for improvement
here, such as a method that takes 2 points in time
and gives you a more precise summarization of the duration
in terms of actual dates.
(And we will do this in a followup)

To avoid changing behaviour where this method is called
the default interval of `month` is not specified, so
existing calls return as they did before.
For callers that want `month` granularity, they can
specify month.

Bug: T219397
Change-Id: I5a07e14629fd9f67b1f17df3db75b68f98295dfb