foreign-resources: Add purl field
This is mainly for the benefit of the CycloneDX export, which
until now did not contain anything that would identify the package
in a machine-readable way.

Also add purl codes to existing packages, based on their current
download location.

Also fix the version declaration for one package where it
mismatched the download URL.

Change-Id: Ib37afdba44b069792e5df7bebb74cf5f588148c4
Bug: T363589