Fix hashes (#) in "Parser function hooks"
To display the names as they're used in wikitext, we must look up the
right syntax, as not all of them are prefixed with "#".

This also displays them in content language (not sure if that's good,
but if we want them to look like they're supposed to be used in
wikitext, I guess it is?), in the preferred case, and in the right
order (order of hooks like 'lc' and 'lcfirst' was previously
reversed, because the trailing '}}' was considered when sorting).

Follow-up to 9308818ee9d70996ac778d6374d3083c0646a20f.

Change-Id: Ia7c505e196e7ae28badc6314b2d6c91f8be45689