Make EncryptedPassword work with Argon2Password
Currently EncryptPassword doesn't work, because it just uses ->verify()
from the Password class. That method does just regenerate the hash of
the password provided and then checks whether they match.

The problem is that in the case of Argon2 being the underlying hashing
algorithm the salt is not known and PHP is just going to generate a new
one while regenerating the hash (in order to compare it then) and thus
the hashes are not going to match making logins fail.

So instead lets just decrypt the stored password hash and then use the
corresponding ->verify() method from the underlying Password type.

Bug: T150647
Bug: T216682
Change-Id: I053aade9dcaf156a26f2f6bebd17773c4df30b9c
(cherry picked from commit 763c5cdfd5138a41ef0cc8ffadbfa247663ace08)