pinia: add vue-demi module to fix dev mode HMR
Add vue-demi as a ResourceLoader foreign resource. dc17d0e1e428 added
the pinia module to core, loading Vue in place of vue-demi. This worked
because pinia only uses vue-demi functions behind `isVue2` checks (which
do not actually exist on the Vue object but just evaluate to falsy),
except for when using hot module replacement in the development mode of
pinia.

Another way to fix pinia calling missing vue-demi functions is by adding
vue-demi as a RL module and adding it as a dependency of pinia, but this
would cause an extra HTTP request for a 560 byte file even in
production, where it is not used.

Bug: T364518
Change-Id: Ib27ac4550e4348c399dfc7e547982c1bf416a1b7