resources: Use Codex tokens for background of mw-datatable
Make the on-hover row highlighting colors use Codex tokens instead of
hardcoded colors. This also helps with night mode support in special
pages and other places that use the mw-datatable class, and can serve as
a starting point for on-wiki templates that emulate this behaviour (such
as https://en.wikipedia.org/wiki/Template:Row_hover_highlight, and many
others in more languages).

This patch does not touch the table headers, as they use a color
(#eaeeff) that does not exist as a Codex token, and would therefore have
to be changed.

Related tasks, but not (partially or in full) resolved by this patch:
T214208, T361048.

Change-Id: I1b2f83fc6b6b1238d68035d58493ca41a2adf252