rdbms: Tighten type checks and PHPDoc for SQL WHERE/$conds arrays
In all cases, these mistakes would have already caused an error
("Wikimedia\Rdbms\Expression could not be converted to string")
or a notice ("Array to string conversion"). But now we also get
Phan warnings to avoid them, and a better message if they happen.

Add tests describing the invalid cases.

Change-Id: Idd4ba30a8145451e85018c35d5441910af76f009