LinkFilter::makeLikeArray: Fix another 'path' access
If a news: or mailto: URL is specified with two slashes, it will have a
'host' rather than a 'path' after all, so this workaround is unnecessary
and should be skipped in that case; compare also change Idc6b389da9
(commit ec1b572362) for makeIndexes().

I’m not very sure that the test case makes much sense, but it’s at least
enough to trigger the error and verify the fix.

Bug: T364743
Change-Id: I09be813e661b80968da00d8a898b2add8c95fec7