Do not mark HttpRequestFactory as @internal
The @internal annotation was added
in  I3e5e54afa21f8c099877b3765c769f2d2f8f5d61
with no explanation. The current annotations advise extensions
to not use HttpRequestFactory, which renders extensions
unable to make HTTP requests when they need to.

This appears to be a declaration added by a mistake,
intending to make the constructor itself as @internal.
Since constructors are unstable by default, this patch
removes the @internal annotation with no replacement.

Change-Id: I9696034fb014c6a87b4532bff22d8c9d494e9bb1