mediawiki.router: Document as module instead of public class
This follows-up I03750273f, which moved OO.Router into MediaWiki core,
however at the same time also added it a seemingly new and public class
to the JS API documentation. This was not previously provided/promised
by MW itself, but rather a contract between oojs-router and MW core
to fulfil mediawiki.router, where semver-major is easy to handle since
it allows for a place to reconcile differences.

Now that it is in core, this extra layer is not needed. We can keep it
as a class for now, as that might make future testing easier (it lacks
unit tests at the moment), but we don't need to expose end-users
of mediawiki.router to that detail.

Bug: T358813
Change-Id: Id88e218f3c3e34c89dc205134e038ea54d16d8f4