Drop call to supportsMenu
The supportsMenu is intended for checking if a skin renders
a menu with a given ID.

There is no menu
in any skin with the element ID p-special-contribute so this
will always return false and using it this way was not how it was
intended to be used.

It seems the desire here was for more control for skins to opt in
but given the feature flag was meant to be temporary
I think the discussion that led to this in T319327#8379688 was
premature.

The config check should be enough here - the check is only
used to determine whether a Skin should link to Special:Contribute
or Special:Contributions in its menus and hopefully soon it will always
be true - this is not a skin-specific feature!

Also improve documentation to make the use of Skin::supportsMenu
clearer.

Follow up to I67779b88c4904462f9d847b3edb08a65306b2bfb

Change-Id: I41d25fd59bc2ee01494376c096f94f9a6f3b968b