Simplify RCFilters layout using flexbox
Compared to the previous approach using table layouts, it is much
easier to control which elements are allowed to extend to maximum
available width, and to provide alternative layouts for mobile.
There should be no visual changes.

It doesn't require using `white-space: nowrap` to prevent elements
from being squished, which will help with T237244.

Bug: T237244
Change-Id: Ia267e94a435f9849e5eeb7fe689c93cf0561903b