extension registration: Fix handling of null default values
The current implementation of ExtensionRegistration violates the
documented behavior for extension registration, as described in
[[mw:Manual:Extension.json/Schema]]. According to the documentation,
when at least one of the global value and the default value is not an
array, the merge strategy is ignored and the global value will simply
override the default. However, an optimization incorrectly replaces a
configured empty array with the default null value.

The issue was fixed by applying the optimization after checking the
violated condition. Tests have been expanded for similar cases and
now additionally provide array diffs on failure.

WMF production configurations were evaluated with and without the
patch, and differences were analyzed on a per-wiki basis. This patch
results in two changes:

1. $wgCirrusSearchWriteClusters changes from null to []. This is the
 desired behaviour.

2. $wgLiqudThreadsAllowUserControlNamespaces changes from null to [].
 null means "no limits" and an empty array means "nothing allowed".
 This matches the intended configuration and will be corrected
 by this patch.

Bug: T365190
Change-Id: Ia3eaf6a607840e3e3134bded37e45050b7f247bf