Clear $wgTitle state before suite runs
`MessageCache::parse` pulls in global `$wgTitle` on line 1492 and
sets the page for the parse to this value of none is set. Some
tests (for example SummaryFormatterTest in Wikibase that extends
MediaWikiLangTestCase) interact with the MessageCache in contexts
where the page is not set, making the test unreliable if other
test cases have run beforehand and set `$wgTitle`.

Bug: T365130
Change-Id: Ib4ee7dc5102d5998aa65dcafb4efa531c2533c69