REST: support rest.php when another RestPath is configured.
When a "pretty" RestPath is configured for outside requests, we still
need the "ugly" path to work for backwards compatibility and internal
requests.

For example, the REST API could be available through /api/ as
well as /w/rest.php. This is equivalent to supporting page views
at /wiki/ as well as /w/index.php.

We can already set RestPath to /api, but then requests to /w/rest.php would fail. We need a way to support both at once.

Change-Id: I9ff8e7345e88b1a4235b6870ec04a87527ed271f