rdbms: Add Phan type hint for allowed operators in expressions
It is really annoying that we have to escape '<' and '>'
as '\x3E' and '\x3C', otherwise Phan will not accept them
(emitting PhanUnextractableAnnotation).

Phan will also escape them in the output, generating errors like:
  Argument 2 ($op) is '=<' of type '=\x3c'
  but \Wikimedia\Rdbms\IDatabase::expr() takes
  '!='|'='|'LIKE'|'NOT LIKE'|'\x3c'|'\x3c='|'\x3e'|'\x3e='

I've wanted to do this for a while, but I wasn't sure if it was
worth it, because the escaping makes the type hints and the errors
hard to read. Today I noticed a bug in UserSelectQueryBuilder
that would have been caught by this, and convinced myself.

Change-Id: I7a72368446f463a99b3d0cc7a90e8f7cdca454fe