Skin: Responsive images is a core feature for responsive skins
This reattempts commit 172e8d942a8157c9b752de05b3101666c30ba65d.
The change is now scoped to the presence of the skin--responsive class
to facilitate further testing and a more cautious roll out. The
skin--responsive class is only added to responsive skins (currently
Minerva, Timeless, Monobook) which already have this style.

The styles are forward thinking and will not work if
$wgParserEnableLegacyMediaDOM and $wgUseLegacyMediaStyles
is enabled.

Bug: T113101
Change-Id: I604ea436a816939cd6b2d9ef8c352e319ffcbec0