jquery.makeCollapsible: fallback to original collapsible
610b9ac91c4b changed the behaviour of jquery.makeCollapsible to check if
the toggle is inside the original collapsible. If not, it searches up the DOM tree to find a .mw-collapsible element. However, this new behaviour causes jquery.makeCollapsible to do nothing if the toggle link has been moved outside of the collapsible but not in the children of the
collapsible.

Add a fallback that uses the original collapsible in case one could not
be found.

Bug: T364712
Change-Id: I6ace62acccdbdd1fbc6307f2090d68f601477a6c