ServiceWiring: Drop `_LocalClusterCache` service
With the ObjectCacheFactory refactor in T358346, this service is
pretty much used in 2 places, (1) service wiring which can be replaced
and (2) in MediaWikiIntegrationTestCase.

The service can be removed entirely and the local cluster cache gotten
from the ObjectCacheFactory service injected via service wiring. Hence,
usage of this internal service can be removed.

NOTE: See followup: 7475063 for the TODO to drop this service entirely.

Change-Id: I173f04456df6cc1782cec2f16d6653c6b0826dce